requirements.txt
seaborn~=0.11.2
numpy~=1.22.3
matplotlib~=3.5.1
pyjnius~=1.4.1